requirements.txt
rumps>=0.4.0
pynput>=1.7.6
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.10.0
elevenlabs>=1.0.0
pyperclip>=1.8.2
pyautogui>=0.9.54
pyobjc-framework-Cocoa>=9.0
pyobjc-framework-Quartz>=9.0
keyring>=24.0.0
Pillow>=10.0.0
